+++ src/Milestones/Block.php
@@ -50,6 +50,7 @@ class Block {
 				'title'       => $attributes['title'],
 				'description' => $attributes['description'],
 				'image'       => $attributes['image'],
+				'ids'         => $attributes['ids'],
 			]
 		);
 		return $milestone->getOutput();